msm8610: Fix CTS issue in libcore.package.okhttp.
The test fails only on 8610 512 MB devices
due to low memory. Through debugging, it was
found that the free memory and heap growth
limit were very low compared to 8226 targets.
Note that this test was not present in KK,
but introduced in L.

Since L requires more memory than KK, the fix
is to increase heapgrowthlimit from 48m to 96m.

However, for L, 512 MB 8610 devices are not going
to be used. Instead, 1GB devices will be used. Even
then, this change is still required. In addition
to this, we need to set other parameters to
appropriate 1 GB related settings.

CRs-fixed: 749728

Change-Id: I8e612a1477645a666e8483d540c82870222dd22e